Rename a file and move it to another folder
