+++ src/Mathematics.NET/LinearAlgebra/LinAlgExtensions.cs
@@ -31,6 +31,7 @@ using System.Runtime.CompilerServices;
 using System.Runtime.Intrinsics;
 using System.Text;
 using Mathematics.NET.AutoDiff;
+using Mathematics.NET.Utilities;
 
 namespace Mathematics.NET.LinearAlgebra;
